1) rename tests to "circular dependency 1...5" 2) move test to separate packages like architecture1..5
